moved loading and error to common
