--- src/features/Loading/styled.js
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
