+++ src/features/Movies/MovieList/index.js
@@ -2,8 +2,8 @@ import { useDispatch, useSelector } from "react-redux";
 import { fetchMovies, selectMovies, selectLoading } from "./moviesSlice";
 import { useEffect } from "react";
 import Success from "./Success";
-import Loading from "../../Loading";
-import Error from "../../Error";
+import Loading from "../../../common/Loading";
+import Error from "../../../common/Error";
 
 const MovieList = () => {
   const dispatch = useDispatch();
